pyproject.toml
[project]
name = "marketing-material-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.128.0",
    "jsonschema>=4.25.1",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pymatting>=1.1.14",
    "python-multipart>=0.0.21",
    "rembg>=2.0.69",
    "reportlab>=4.4.7",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
]
